Refactor out manager construction
