+++ supported_manager.c
@@ -5,6 +5,8 @@
 #include "flatpak.h"
 #include "guix.h"
 #include "snap.h"
+#include "package_manager.h"
+#include "requested_action.h"
 
 enum SupportedManager getManagerIndex(char* managerName) {
     if (strcmp(managerName, APT) == 0) {
@@ -20,5 +22,25 @@ enum SupportedManager getManagerIndex(char* managerName) {
     } else {
         return InvalidManager;
     }
+}
+
+PackageManager* apt(RequestedAction* requestedAction) {
+    return definePackageManager(APT, getAptCommands(getRequestedActionTarget(requestedAction)), getApt(getRequestedActionManagers(requestedAction)));
+}
+
+PackageManager* brew(RequestedAction* requestedAction) {
+    return definePackageManager(BREW, getBrewCommands(getRequestedActionTarget(requestedAction)), getBrew(getRequestedActionManagers(requestedAction)));
+}
+
+PackageManager* flatpak(RequestedAction* requestedAction) {
+    return definePackageManager(FLATPAK, getFlatpakCommands(getRequestedActionTarget(requestedAction)), getFlatpak(getRequestedActionManagers(requestedAction)));
+}
+
+PackageManager* guix(RequestedAction* requestedAction) {
+    return definePackageManager(GUIX, getGuixCommands(getRequestedActionTarget(requestedAction)), getGuix(getRequestedActionManagers(requestedAction)));
+}
+
+PackageManager* snap(RequestedAction* requestedAction) {
+    return definePackageManager(SNAP, getSnapCommands(getRequestedActionTarget(requestedAction)), getSnap(getRequestedActionManagers(requestedAction)));
 }
